filteredRecipeData.ts to clean recipeData.ts
